Checkpoint: separate component for dropdown
